refactor(frontend): consolidate purchase orders workspace hook
