+++ frontend/src/pages/purchasing/PurchaseOrdersPage.tsx
@@ -1,27 +1,17 @@
-import React, { useCallback, useMemo } from 'react'
-import { useNavigate, useSearchParams } from 'react-router-dom'
+import React, { useCallback, useMemo, useState } from 'react'
+import { useNavigate } from 'react-router-dom'
 
 import GenericListPage from '@/components/common/GenericListPage'
 import PurchaseOrderContextHeader from './components/PurchaseOrderContextHeader'
 import PurchaseOrdersDialogs from './components/PurchaseOrdersDialogs'
 import PurchaseOrdersTable from './components/PurchaseOrdersTable'
 import PurchaseOrderWorkspaceCard from './components/PurchaseOrderWorkspaceCard'
-import { usePurchaseOrdersActions } from './hooks/purchaseOrdersActions'
-import { usePurchaseOrdersPageState } from './hooks/purchaseOrdersPageState'
-import { usePurchaseOrdersSelection } from './hooks/purchaseOrdersSelection'
+import { usePurchaseOrdersWorkspace } from './hooks/usePurchaseOrdersWorkspace'
 
 import { useFilterBar } from '@/hooks/useFilterBar'
-import { useNotification } from '@/hooks/useNotification'
-import { useKeyboardShortcuts } from '@/hooks/useSearchAndFilter'
 import { useAppDispatch, useAppSelector } from '@/hooks/useRedux'
 import {
-  useDeletePurchaseOrderMutation,
   useGetPurchaseOrdersQuery,
-  useLazyGetPurchaseOrderQuery,
-  useMarkPurchaseOrderAsUnpaidMutation,
-  useReceiveGoodsMutation,
-  useRecordOrderPaymentsMutation,
-  useReturnGoodsMutation,
 } from '@/store/api/purchasingApi'
 import { selectSelectedPurchaseOrder } from '@/store/slices/purchasingSlice'
 import type { FilterBarConfig, PeriodValue } from '@/types/filterBar.types'
@@ -38,10 +28,11 @@ interface PurchaseOrderFilters {
 export const PurchaseOrdersPage: React.FC = () => {
   const navigate = useNavigate()
   const dispatch = useAppDispatch()
-  const { showSuccess, showError } = useNotification()
-  const [searchParams, setSearchParams] = useSearchParams()
-  const pageState = usePurchaseOrdersPageState()
   const selectedOrder = useAppSelector(selectSelectedPurchaseOrder)
+  const [sorting, setSorting] = useState<{ sortBy: string; sortOrder: 'asc' | 'desc' }>({
+    sortBy: 'orderNumber',
+    sortOrder: 'asc',
+  })
 
   const filterConfig = useMemo<FilterBarConfig<PurchaseOrderFilters>>(
     () => ({
@@ -95,15 +86,15 @@ export const PurchaseOrdersPage: React.FC = () => {
   }, [filterBar.appliedFilters.period, weekStartsOn])
 
   const queryParams = useMemo(() => ({
-    sortBy: pageState.sorting.sortBy,
-    sortOrder: pageState.sorting.sortOrder.toUpperCase(),
+    sortBy: sorting.sortBy,
+    sortOrder: sorting.sortOrder.toUpperCase(),
     search: filterBar.appliedFilters.search || undefined,
     supplierId: filterBar.appliedFilters.supplierId || undefined,
     paymentStatus: filterBar.appliedFilters.paymentStatus || undefined,
     status: filterBar.appliedFilters.status || undefined,
     orderDateFrom: dateRange.fromDate,
     orderDateTo: dateRange.toDate,
-  }), [dateRange, filterBar.appliedFilters, pageState.sorting.sortBy, pageState.sorting.sortOrder])
+  }), [dateRange, filterBar.appliedFilters, sorting.sortBy, sorting.sortOrder])
 
   const {
     data: purchaseOrdersResponse,
@@ -111,12 +102,6 @@ export const PurchaseOrdersPage: React.FC = () => {
     error: purchaseOrdersError,
     refetch: refetchOrders,
   } = useGetPurchaseOrdersQuery(queryParams)
-  const [fetchPurchaseOrder] = useLazyGetPurchaseOrderQuery()
-  const [receiveGoods] = useReceiveGoodsMutation()
-  const [returnGoods] = useReturnGoodsMutation()
-  const [markPurchaseOrderAsUnpaid] = useMarkPurchaseOrderAsUnpaidMutation()
-  const [recordOrderPayments] = useRecordOrderPaymentsMutation()
-  const [deletePurchaseOrder] = useDeletePurchaseOrderMutation()
 
   const purchaseOrders = purchaseOrdersResponse?.data || []
   const pagination = purchaseOrdersResponse?.meta
@@ -128,93 +113,36 @@ export const PurchaseOrdersPage: React.FC = () => {
   const loadOrders = useCallback(() => {
     void refetchOrders()
   }, [refetchOrders])
-
-  const selection = usePurchaseOrdersSelection({
+  const workspace = usePurchaseOrdersWorkspace({
     dispatch,
     purchaseOrders,
     selectedOrder,
-    focusedOrderIndex: pageState.focusedOrderIndex,
-    setFocusedOrderIndex: pageState.setFocusedOrderIndex,
-    searchParams,
-    setSearchParams,
-    fetchPurchaseOrder,
-    pendingHighlightId: pageState.pendingHighlightId,
-    setPendingHighlightId: pageState.setPendingHighlightId,
-    orderListRef: pageState.orderListRef,
-    searchInputRef: pageState.searchInputRef,
-    processedHighlightRef: pageState.processedHighlightRef,
-    userHasNavigatedRef: pageState.userHasNavigatedRef,
-    setJournalEntryRef: pageState.setJournalEntryRef,
-    setJournalEntryRefLoading: pageState.setJournalEntryRefLoading,
-  })
-
-  const actions = usePurchaseOrdersActions({
-    dispatch,
-    navigate,
-    purchaseOrders,
-    selectedOrder,
-    receiveGoods,
-    returnGoods,
-    markPurchaseOrderAsUnpaid,
-    recordOrderPayments,
-    deletePurchaseOrder,
-    loadOrders,
-    showSuccess,
-    showError,
-    setBlockedDialogType: pageState.setBlockedDialogType,
-    setBlockedDialogOpen: pageState.setBlockedDialogOpen,
-    setIsLoading: pageState.setIsLoading,
-    setPaymentDialogOrder: pageState.setPaymentDialogOrder,
-    setPaymentDialogOpen: pageState.setPaymentDialogOpen,
-    setDeleteConfirmOpen: pageState.setDeleteConfirmOpen,
-    setOrderToDelete: pageState.setOrderToDelete,
-    setFocusedOrderIndex: pageState.setFocusedOrderIndex,
+    refetchOrders: loadOrders,
   })
 
   const handleSort = useCallback((field: string) => {
-    pageState.setSorting((prev) => ({
+    setSorting((prev) => ({
       ...prev,
       sortBy: field,
       sortOrder: prev.sortBy === field && prev.sortOrder === 'desc' ? 'asc' : 'desc',
     }))
-  }, [pageState])
-
-  useKeyboardShortcuts({
-    onSearch: selection.focusSearchInput,
-    onArrowUp: selection.handleNavigateUp,
-    onArrowDown: selection.handleNavigateDown,
-  })
-
-  const navigateToGoodsReceived = useCallback((grnId: string) => {
-    navigate(`/purchasing/goods-received?grnId=${grnId}`)
-  }, [navigate])
-
-  const navigateToVendorPayment = useCallback((paymentId: string) => {
-    navigate(`/purchasing/vendor-payments?vpId=${paymentId}`)
-  }, [navigate])
-
-  const navigateToJournalEntry = useCallback(() => {
-    if (!pageState.journalEntryRef) return
-    navigate(
-      `/accounting/journal-entries?sourceType=${pageState.journalEntryRef.sourceType}&sourceId=${pageState.journalEntryRef.sourceId}`,
-    )
-  }, [navigate, pageState.journalEntryRef])
+  }, [])
 
   return (
     <GenericListPage
       title="Purchase Orders"
       subtitle="Manage supplier purchase orders and procurement"
-      secondaryAction={{ label: 'View Deleted', onClick: () => pageState.setDeletedOrdersDialogOpen(true) }}
+      secondaryAction={{ label: 'View Deleted', onClick: () => workspace.setDeletedOrdersDialogOpen(true) }}
       primaryAction={{ label: 'Create Order', onClick: () => navigate('/purchasing/orders/create') }}
       filterConfig={filterConfig}
       draftFilters={filterBar.draftFilters}
       handlers={filterBar.handlers}
       hasActiveFilters={filterBar.hasActiveFilters}
-      searchInputRef={pageState.searchInputRef}
+      searchInputRef={workspace.searchInputRef}
       sort={{
         field: 'orderNumber',
-        sortBy: pageState.sorting.sortBy,
-        sortOrder: pageState.sorting.sortOrder,
+        sortBy: sorting.sortBy,
+        sortOrder: sorting.sortOrder,
         onSort: handleSort,
       }}
       error={error}
@@ -224,58 +152,58 @@ export const PurchaseOrdersPage: React.FC = () => {
           loading={loading}
           total={pagination?.total || 0}
           selectedOrderId={selectedOrder?.id}
-          focusedOrderIndex={pageState.focusedOrderIndex}
-          onOrderSelect={selection.handleOrderSelect}
-          orderListRef={pageState.orderListRef}
+          focusedOrderIndex={workspace.focusedOrderIndex}
+          onOrderSelect={workspace.handleOrderSelect}
+          orderListRef={workspace.orderListRef}
         />
       )}
       headerSlot={(
         <PurchaseOrderContextHeader
           selectedOrder={selectedOrder}
-          isLoading={pageState.isLoading}
-          journalEntryRef={pageState.journalEntryRef}
-          journalEntryRefLoading={pageState.journalEntryRefLoading}
-          onEditClick={actions.handleEditClick}
-          onDeleteClick={actions.handleDeleteClick}
-          onPrint={() => pageState.setPrintDialogOpen(true)}
-          onNavigateToGoodsReceived={navigateToGoodsReceived}
-          onNavigateToVendorPayment={navigateToVendorPayment}
-          onNavigateToJournalEntry={navigateToJournalEntry}
-          onUnpay={actions.handleUnpay}
-          onOpenPaymentDialog={actions.handleOpenPaymentDialog}
-          onReturn={actions.handleReturn}
-          onReceive={actions.handleReceive}
+          isLoading={workspace.isLoading}
+          journalEntryRef={workspace.journalEntryRef}
+          journalEntryRefLoading={workspace.journalEntryRefLoading}
+          onEditClick={workspace.handleEditClick}
+          onDeleteClick={workspace.handleDeleteClick}
+          onPrint={() => workspace.setPrintDialogOpen(true)}
+          onNavigateToGoodsReceived={workspace.navigateToGoodsReceived}
+          onNavigateToVendorPayment={workspace.navigateToVendorPayment}
+          onNavigateToJournalEntry={workspace.navigateToJournalEntry}
+          onUnpay={workspace.handleUnpay}
+          onOpenPaymentDialog={workspace.handleOpenPaymentDialog}
+          onReturn={workspace.handleReturn}
+          onReceive={workspace.handleReceive}
         />
       )}
       workspaceSlot={<PurchaseOrderWorkspaceCard selectedOrder={selectedOrder} />}
       dialogs={(
         <PurchaseOrdersDialogs
           selectedOrder={selectedOrder}
-          deleteConfirmOpen={pageState.deleteConfirmOpen}
-          orderToDelete={pageState.orderToDelete}
+          deleteConfirmOpen={workspace.deleteConfirmOpen}
+          orderToDelete={workspace.orderToDelete}
           onCancelDelete={() => {
-            pageState.setDeleteConfirmOpen(false)
-            pageState.setOrderToDelete(null)
+            workspace.setDeleteConfirmOpen(false)
+            workspace.setOrderToDelete(null)
           }}
-          onConfirmDelete={() => actions.handleDeleteConfirm(pageState.orderToDelete)}
-          deletedOrdersDialogOpen={pageState.deletedOrdersDialogOpen}
-          onCloseDeletedOrdersDialog={() => pageState.setDeletedOrdersDialogOpen(false)}
+          onConfirmDelete={() => workspace.handleDeleteConfirm(workspace.orderToDelete)}
+          deletedOrdersDialogOpen={workspace.deletedOrdersDialogOpen}
+          onCloseDeletedOrdersDialog={() => workspace.setDeletedOrdersDialogOpen(false)}
           onRefreshDeletedOrders={loadOrders}
-          blockedDialogOpen={pageState.blockedDialogOpen}
-          blockedDialogType={pageState.blockedDialogType}
-          onCloseBlockedDialog={() => pageState.setBlockedDialogOpen(false)}
-          onReturnAndEdit={actions.handleReturnAndEdit}
-          onReturnOnly={actions.handleReturnOnly}
-          onUnpayAndEdit={actions.handleUnpayAndEdit}
-          onReturnAndDelete={actions.handleReturnAndDelete}
-          onUnpayAndDelete={actions.handleUnpayAndDelete}
-          isLoading={pageState.isLoading}
-          printDialogOpen={pageState.printDialogOpen}
-          onClosePrintDialog={() => pageState.setPrintDialogOpen(false)}
-          paymentDialogOpen={pageState.paymentDialogOpen}
-          paymentDialogOrder={pageState.paymentDialogOrder}
-          onClosePaymentDialog={() => pageState.setPaymentDialogOpen(false)}
-          onSubmitPayments={actions.handleRecordPayments}
+          blockedDialogOpen={workspace.blockedDialogOpen}
+          blockedDialogType={workspace.blockedDialogType}
+          onCloseBlockedDialog={() => workspace.setBlockedDialogOpen(false)}
+          onReturnAndEdit={workspace.handleReturnAndEdit}
+          onReturnOnly={workspace.handleReturnOnly}
+          onUnpayAndEdit={workspace.handleUnpayAndEdit}
+          onReturnAndDelete={workspace.handleReturnAndDelete}
+          onUnpayAndDelete={workspace.handleUnpayAndDelete}
+          isLoading={workspace.isLoading}
+          printDialogOpen={workspace.printDialogOpen}
+          onClosePrintDialog={() => workspace.setPrintDialogOpen(false)}
+          paymentDialogOpen={workspace.paymentDialogOpen}
+          paymentDialogOrder={workspace.paymentDialogOrder}
+          onClosePaymentDialog={() => workspace.setPaymentDialogOpen(false)}
+          onSubmitPayments={workspace.handleRecordPayments}
         />
       )}
     />
