+++ frontend/src/pages/purchasing/__tests__/PurchaseOrdersPage.filterbar.test.tsx
@@ -56,33 +56,51 @@ vi.mock('../components/PurchaseOrderContextHeader', () => ({ default: () => <div
 vi.mock('../components/PurchaseOrdersTable', () => ({ default: () => <div>PurchaseOrdersTable</div> }))
 vi.mock('../components/PurchaseOrderWorkspaceCard', () => ({ default: () => <div>PurchaseOrderWorkspaceCard</div> }))
 vi.mock('../components/PurchaseOrdersDialogs', () => ({ default: () => <div>PurchaseOrdersDialogs</div> }))
-vi.mock('../hooks/purchaseOrdersActions', () => ({
-  usePurchaseOrdersActions: () => ({
-    handleEditClick: vi.fn(),
-    handleDeleteClick: vi.fn(),
-    handleUnpay: vi.fn(),
-    handleOpenPaymentDialog: vi.fn(),
-    handleReturn: vi.fn(),
+vi.mock('../hooks/usePurchaseOrdersWorkspace', () => ({
+  usePurchaseOrdersWorkspace: () => ({
+    sorting: { sortBy: 'orderNumber', sortOrder: 'asc' },
+    handleSort: vi.fn(),
+    focusedOrderIndex: -1,
+    deleteConfirmOpen: false,
+    setDeleteConfirmOpen: vi.fn(),
+    orderToDelete: null,
+    setOrderToDelete: vi.fn(),
+    deletedOrdersDialogOpen: false,
+    setDeletedOrdersDialogOpen: vi.fn(),
+    blockedDialogOpen: false,
+    setBlockedDialogOpen: vi.fn(),
+    printDialogOpen: false,
+    setPrintDialogOpen: vi.fn(),
+    blockedDialogType: 'edit',
+    isLoading: false,
+    paymentDialogOpen: false,
+    setPaymentDialogOpen: vi.fn(),
+    paymentDialogOrder: null,
+    journalEntryRef: null,
+    journalEntryRefLoading: false,
+    orderListRef: { current: null },
+    searchInputRef: { current: null },
+    handleOrderSelect: vi.fn(),
+    handleNavigateUp: vi.fn(),
+    handleNavigateDown: vi.fn(),
+    focusSearchInput: vi.fn(),
     handleReceive: vi.fn(),
-    handleDeleteConfirm: vi.fn(),
+    handleReturn: vi.fn(),
+    handleEditClick: vi.fn(),
     handleReturnAndEdit: vi.fn(),
     handleReturnOnly: vi.fn(),
     handleUnpayAndEdit: vi.fn(),
     handleReturnAndDelete: vi.fn(),
     handleUnpayAndDelete: vi.fn(),
+    handleUnpay: vi.fn(),
+    handleOpenPaymentDialog: vi.fn(),
     handleRecordPayments: vi.fn(),
+    handleDeleteClick: vi.fn(),
+    handleDeleteConfirm: vi.fn(),
+    navigateToGoodsReceived: vi.fn(),
+    navigateToVendorPayment: vi.fn(),
+    navigateToJournalEntry: vi.fn(),
   }),
-}))
-vi.mock('../hooks/purchaseOrdersSelection', () => ({
-  usePurchaseOrdersSelection: () => ({
-    handleOrderSelect: vi.fn(),
-    handleNavigateUp: vi.fn(),
-    handleNavigateDown: vi.fn(),
-    focusSearchInput: vi.fn(),
-  }),
-}))
-vi.mock('@/hooks/useNotification', () => ({
-  useNotification: () => ({ showSuccess: vi.fn(), showError: vi.fn() }),
 }))
 
 function renderPage(initialUrl = '/') {
